Need to gracefully handle 404
If we get a 404 when attempting to read a resource, we need to just
indicate that the resource was deleted by setting the ID to empty
and just return without an error. This is so that terraform can refresh
it's state if the resource has been delete from the provider.