Move schema definition for connection speed
Move the schema definitions for connection speed in to the individual
resources since each provider has different requirements on when the
connection would need to be recreated vs just being updated.

Also added a unit test to make sure that when the speed is updated for
AWS connections, we are actually creating a new connection.