Update documentation and integrate with terraform-website repo
The terraform-websitre repo handles the automated deployment and rendering of the provide layout and
documentation markdown files for our data source and resources. Updated the documentation so it renders
similiarly to other providers. Note! there are some changes required to the terraform-websie to get
the website make target working correctly. These will have to be merged once the provider is published
in the terraform-providers repo in github.