Fix bug in acceptance tests
Since we were setting a default API URL, we weren't reading it from the environment
like we were before. Now we try to read the values from the environment and set these
as the default. If nothing was found, we leave them as nil. That way the Pureport SDK
can retrieve the values from other sources.