Release 0.2.0 (#14)
* Name wasn't being correctly generated for master branch

* Start adding Documentation for Data Sources & Resources

* Fill out the documentation for Data Sources and Resources

* Add some additional test and lint targets

* Update documentation and integrate with terraform-website repo

The terraform-websitre repo handles the automated deployment and rendering of the provide layout and
documentation markdown files for our data source and resources. Updated the documentation so it renders
similiarly to other providers. Note! there are some changes required to the terraform-websie to get
the website make target working correctly. These will have to be merged once the provider is published
in the terraform-providers repo in github.

* Documentation updates from code review

* Add support for reading authetication information from provider config

The standard pattern for setting authentication for a provider is by configuring
it through the provider block of the terrafrom configuration. This change should add
that for our provider.

* Add configuration for the API Endpoint

Since we can test against multiple endpoints, add a provider configuration option
to point to the one we want.

* Rename api_endpoint to api_url

* Set the SDK Log level to the same as Terraform

* Update golang dependencies including new Pureport SDK

* Fix bug in acceptance tests

Since we were setting a default API URL, we weren't reading it from the environment
like we were before. Now we try to read the values from the environment and set these
as the default. If nothing was found, we leave them as nil. That way the Pureport SDK
can retrieve the values from other sources.

* Add documenation for the Provider configurations

* Update Pureport SDK

* Update the version number

* Update to release version of Pureport SDK

* Add StandardGateway remoteId attribute (#15)

For AWS, Azure, and Google, we need to provide a RemoteID in the gateway attributes
for the remote resource that is created at the Cloud Provider.

* Feature/fix code review comments (#16)

* Add StandardGateway remoteId attribute

For AWS, Azure, and Google, we need to provide a RemoteID in the gateway attributes
for the remote resource that is created at the Cloud Provider.

* Fix typos and grammer

* Fix Resource descriptions

* Remove redundant Connection