Fixes #19 Crash when parsing IKE Config (#20)
When parsing the IKE config, we were just decoding map[string]interface{}
which is how you would normally define it. But do to the way Terraform
handles nested maps, you have to make them lists. No we decode the
ike_config as a list with one item and ti works as expected.