[PE-1254] CRASH when running acceptance tests (#43)
* Fix crash when running acceptance tests in parallel

When running the acceptance tests in parallel, we were seeing a crash due to concurrent map
modification from the call to logger.SetLevel() from the Pureport SDK. The Pureport SDK is
using go-logger that has a global value for the log level. Since the Terraform acceptance
tests call ProviderConfigure() concurrently, you will see a crash when both threads were attempting
to set the Level.

Adding a mutex to lock the modification of the logger to prevent the crash.

* Fix Log Level mapping between Pureport SDK and TF_LOG

The log levels for Terraform don't exactly line up with each other.
Now we set the appropriate log level for our SDK based on the TF_LOG level
set by the user.