[PE-1054][PE-1056] Add support for tagging resources, filter, and sweep (#29)
* Update Pureport golang SDK to latest

Update the Pureport SDK to the latest version which is using an updated
Swagger interface for creating ports facilities and tagging resources.

* Upgrade to terraform 0.12.2

* Add support for tags

Add support for adding arbitrary tags to Pureport resources.
We should then be able to filter resources based on any applied
tags. This will make it easier to organize and cleanup resources.

* Refactor Acceptance Tests to run in multiple environments

We need to separate out the parameters needed for each of the different
environments we need for deploying our provider.

* Update documentation to include the tags attribute

* Initial implementation of Sweeper

* Change directory structure to match recommendations for multiple environments

* Need to add tags to to the data source connection schemas

* Update to latest Terraform and Azure SDKs

* Update variable documentation

Updating the descriptions for the test-infra variables based on code review
feedback.