[PE-1055] Add support for data source filter schema (#49)
* Update Pureport v1.1.0 golang SDK

Update the Pureport SDK to the latest version which is using an updated
Swagger interface for creating ports facilities and tagging resources.

* Upgrade to terraform 0.12.3

* Add support for tags

Add support for adding arbitrary tags to Pureport resources.
We should then be able to filter resources based on any applied
tags. This will make it easier to organize and cleanup resources.

* Refactor Acceptance Tests to run in multiple environments

We need to separate out the parameters needed for each of the different environments we need for deploying our provider.

* Update documentation to include the tags attribute

* Initial implementation of Sweeper

* Change directory structure to match recommendations for multiple environments

* Need to add tags to to the data source connection schemas

* Update to latest Terraform and Azure SDKs

* Update variable documentation

Updating the descriptions for the test-infra variables based on code review
feedback.

* Add filter and schema for filtering individual fields

* Replace name_regex with Name filter

* Add support for nested filter field names

* Fix expected results for dev1

* Rename tests to be more explicit about data source vs. resource

* Only set configuration parameters if they are set

Need to add tags for the resources we want sweep to cleanup.
Also add some debug logging for the sweep tasks.

* Need to comsume sweep flag to prevent reporting error

* Add the missing connection data source documentation

* Removing Gateway Link State

This doesn't look like it was being used for anything and there is a
different configuration parameter for the BGP Status.